DELPHI-227 feat: *create apiPatchAdviceSeeker *add enableAdviceSeekerWalkThrough to appSettings *updated EnableWalkthrough to use both patchConsultant and patchAdviceSeeker apis *add EnableWalkthrough to the adviceSeekers profile *add initialEnquirySent to the UserDataInterface *add new PatchAdviceSeekerDTO interface *Change Walkthrough to manage the old consultant and the new AS Walkthrough *Rename old Walkthrough to WalkthroughConsultant
