+++ src/globalState/interfaces/AppConfig/AppSettingsInterface.ts
@@ -9,6 +9,8 @@ export interface AppSettingsInterface {
 	calcomUrl?: string;
 	/** Feature flag to enable walkthrough (false by default here & true in the theme repo) */
 	enableWalkthrough?: boolean;
+	/** Feature flag to enable the advice seeker walkthrough (false by default here & true in the theme repo) */
+	enableAdviceSeekerWalkThrough?: boolean;
 	/** Feature flag to enable Video-Termine page */
 	disableVideoAppointments?: boolean;
 	/** Feature flag to enable the multi tenancy with a single domain ex: lands */
